mcp_daemon::error::types
Function tool_not_found

pub fn tool_not_found(_name: &str) -> Error

Creates an error for when a requested tool is not found.

This function generates a standardized JSON-RPC error with the METHOD_NOT_FOUND error code and a message indicating that the tool was not found. This should be used when a client attempts to use a tool that doesn’t exist or is not available.

Parameters

name - The name of the tool that was not found

Returns

A JSON-RPC Error object with error code METHOD_NOT_FOUND

Examples

use mcp_daemon::error::tool_not_found;
 
let error = tool_not_found("non_existent_tool");
// Return this error in a response to the client