mcp_daemon::schema
Module types

Modules

builder

Types for composing complex structures.

error

Error types.

Structs

Annotations

Optional annotations for the client. The client can use annotations to inform how objects are used or displayed

AudioContent

Audio provided to or from an LLM.

BlobResourceContents

BlobResourceContents

CallToolRequest

Used by the client to invoke a tool provided by the server.

CallToolRequestParams

CallToolRequestParams

CallToolResult

The server’s response to a tool call.

CancelledNotification

This notification can be sent by either side to indicate that it is cancelling a previously-issued request.

CancelledNotificationParams

CancelledNotificationParams

ClientCapabilities

Capabilities a client may support. Known capabilities are defined here, in this schema, but this is not a closed set: any client can define its own, additional capabilities.

ClientCapabilitiesRoots

Present if the client supports listing roots.

CompleteRequest

A request from the client to the server, to ask for completion options.

CompleteRequestParams

CompleteRequestParams

CompleteRequestParamsArgument

The argument’s information

CompleteResult

The server’s response to a completion/complete request

CompleteResultCompletion

CompleteResultCompletion

CreateMessageRequest

A request from the server to sample an LLM via the client. The client has full discretion over which model to select. The client should also inform the user before beginning sampling, to allow them to inspect the request (human in the loop) and decide whether to approve it.

CreateMessageRequestParams

CreateMessageRequestParams

CreateMessageResult

The client’s response to a sampling/create_message request from the server. The client should inform the user before returning the sampled message, to allow them to inspect the response (human in the loop) and decide whether to allow the server to see it.

Cursor

An opaque token used to represent a cursor for pagination.

EmbeddedResource

The contents of a resource, embedded into a prompt or tool call result.

EmptyResult

EmptyResult

GetPromptRequest

Used by the client to get a prompt provided by the server.

GetPromptRequestParams

GetPromptRequestParams

GetPromptResult

The server’s response to a prompts/get request from the client.

ImageContent

An image provided to or from an LLM.

Implementation

Describes the name and version of an MCP implementation.

InitializeRequest

This request is sent from the client to the server when it first connects, asking it to begin initialization.

InitializeRequestParams

InitializeRequestParams

InitializeResult

After receiving an initialize request from the client, the server sends this response.

InitializedNotification

This notification is sent from the client to the server after initialization has finished.

InitializedNotificationParams

InitializedNotificationParams

JsonrpcBatchRequest

A JSON-RPC batch request, as described in https://www.jsonrpc.org/specification#batch.

JsonrpcBatchResponse

A JSON-RPC batch response, as described in https://www.jsonrpc.org/specification#batch.

JsonrpcError

A response to a request that indicates an error occurred.

JsonrpcErrorError

JsonrpcErrorError

JsonrpcMessage

Refers to any valid JSON-RPC object that can be decoded off the wire, or encoded to be sent.

JsonrpcNotification

A notification which does not expect a response.

JsonrpcNotificationParams

JsonrpcNotificationParams

JsonrpcRequest

A request that expects a response.

JsonrpcRequestParams

JsonrpcRequestParams

JsonrpcRequestParamsMeta

JsonrpcRequestParamsMeta

JsonrpcResponse

A successful (non-error) response to a request.

ListPromptsRequest

Sent from the client to request a list of prompts and prompt templates the server has.

ListPromptsRequestParams

ListPromptsRequestParams

ListPromptsResult

The server’s response to a prompts/list request from the client.

ListResourceTemplatesRequest

Sent from the client to request a list of resource templates the server has.

ListResourceTemplatesRequestParams

ListResourceTemplatesRequestParams

ListResourceTemplatesResult

The server’s response to a resources/templates/list request from the client.

ListResourcesRequest

Sent from the client to request a list of resources the server has.

ListResourcesRequestParams

ListResourcesRequestParams

ListResourcesResult

The server’s response to a resources/list request from the client.

ListRootsRequest

Sent from the server to request a list of root URIs from the client. Roots allow servers to ask for specific directories or files to operate on. A common example for roots is providing a set of repositories or directories a server should operate on.

ListRootsRequestParams

ListRootsRequestParams

ListRootsRequestParamsMeta

ListRootsRequestParamsMeta

ListRootsResult

The client’s response to a roots/list request from the server. This result contains an array of Root objects, each representing a root directory or file that the server can operate on.

ListToolsRequest

Sent from the client to request a list of tools the server has.

ListToolsRequestParams

ListToolsRequestParams

ListToolsResult

The server’s response to a tools/list request from the client.

LoggingMessageNotification

Notification of a log message passed from server to client. If no logging/setLevel request has been sent from the client, the server MAY decide which messages to send automatically.

LoggingMessageNotificationParams

LoggingMessageNotificationParams

ModelHint

Hints to use for model selection.

ModelPreferences

The server’s preferences for model selection, requested of the client during sampling.

Notification

Notification

NotificationParams

NotificationParams

PaginatedRequest

PaginatedRequest

PaginatedRequestParams

PaginatedRequestParams

PaginatedResult

PaginatedResult

PingRequest

A ping, issued by either the server or the client, to check that the other party is still alive. The receiver must promptly respond, or else may be disconnected.

PingRequestParams

PingRequestParams

PingRequestParamsMeta

PingRequestParamsMeta

ProgressNotification

An out-of-band notification used to inform the receiver of a progress update for a long-running request.

ProgressNotificationParams

ProgressNotificationParams

Prompt

A prompt or prompt template that the server offers.

PromptArgument

Describes an argument that a prompt can accept.

PromptListChangedNotification

An optional notification from the server to the client, informing it that the list of prompts it offers has changed. This may be issued by servers without any previous subscription from the client.

PromptListChangedNotificationParams

PromptListChangedNotificationParams

PromptMessage

Describes a message returned as part of a prompt.

PromptReference

Identifies a prompt.

ReadResourceRequest

Sent from the client to the server, to read a specific resource URI.

ReadResourceRequestParams

ReadResourceRequestParams

ReadResourceResult

The server’s response to a resources/read request from the client.

Request

Request

RequestId

RequestParamsMeta

A uniquely identifying ID for a request in JSON-RPC.

Resource

A known resource that the server is capable of reading.

ResourceContents

The contents of a specific resource or sub-resource.

ResourceListChangedNotification

An optional notification from the server to the client, informing it that the list of resources it can read from has changed. This may be issued by servers without any previous subscription from the client.

ResourceListChangedNotificationParams

ResourceListChangedNotificationParams

ResourceReference

A reference to a resource or resource template definition.

ResourceTemplate

A template description for resources available on the server.

ResourceUpdatedNotification

A notification from the server to the client, informing it that a resource has changed and may need to be read again. This should only be sent if the client previously sent a resources/subscribe request.

ResourceUpdatedNotificationParams

ResourceUpdatedNotificationParams

Result

Result

Root

Represents a root directory or file that the server can operate on.

RootsListChangedNotification

A notification from the client to the server, informing it that the list of roots has changed. This notification should be sent whenever the client adds, removes, or modifies any root. The server should then request an updated list of roots using the ListRootsRequest.

RootsListChangedNotificationParams

RootsListChangedNotificationParams

SamplingMessage

Describes a message issued to or received from an LLM API.

ServerCapabilities

Capabilities that a server may support. Known capabilities are defined here, in this schema, but this is not a closed set: any server can define its own, additional capabilities.

ServerCapabilitiesPrompts

Present if the server offers any prompt templates.

ServerCapabilitiesResources

Present if the server offers any resources to read.

ServerCapabilitiesTools

Present if the server offers any tools to call.

SetLevelRequest

A request from the client to the server, to enable or adjust logging.

SetLevelRequestParams

SetLevelRequestParams

SubscribeRequest

Sent from the client to request resources/updated notifications from the server whenever a particular resource changes.

SubscribeRequestParams

SubscribeRequestParams

TextContent

Text provided to or from an LLM.

TextResourceContents

TextResourceContents

Tool

Definition for a tool the client can call.

ToolAnnotations

Additional properties describing a Tool to clients.

ToolInputSchema

A JSON Schema object defining the expected parameters for the tool.

ToolListChangedNotification

An optional notification from the server to the client, informing it that the list of tools it offers has changed. This may be issued by servers without any previous subscription from the client.

ToolListChangedNotificationParams

ToolListChangedNotificationParams

UnsubscribeRequest

Sent from the client to request cancellation of resources/updated notifications from the server. This should follow a previous resources/subscribe request.

UnsubscribeRequestParams

UnsubscribeRequestParams

Enums

CallToolResultContentItem

CallToolResultContentItem

ClientNotification

ClientNotification

ClientRequest

ClientRequest

ClientResult

ClientResult

CompleteRequestParamsRef

CompleteRequestParamsRef

CreateMessageRequestParamsIncludeContext

A request to include context from one or more MCP servers (including the caller), to be attached to the prompt. The client MAY ignore this request.

CreateMessageResultContent

CreateMessageResultContent

EmbeddedResourceResource

EmbeddedResourceResource

JsonrpcBatchRequestItem

JsonrpcBatchRequestItem

JsonrpcBatchResponseItem

JsonrpcBatchResponseItem

JsonrpcMessageSubtype2Item

JsonrpcMessageSubtype2Item

JsonrpcMessageSubtype5Item

JsonrpcMessageSubtype5Item

LoggingLevel

The severity of a log message.

ProgressToken

A progress token, used to associate progress notifications with the original request.

PromptMessageContent

PromptMessageContent

ReadResourceResultContentsItem

ReadResourceResultContentsItem

Role

The sender or recipient of messages and data in a conversation.

SamplingMessageContent

SamplingMessageContent

ServerNotification

ServerNotification

ServerRequest

ServerRequest

ServerResult

ServerResult