mcp_daemon::schema::types
Struct EmbeddedResource

pub struct EmbeddedResource {
    pub annotations: Option<Annotations>,
    pub resource: EmbeddedResourceResource,
    pub type_: String,
}

The contents of a resource, embedded into a prompt or tool call result.

It is up to the client how best to render embedded resources for the benefit of the LLM and/or the user.

JSON schema

{
 "description": "The contents of a resource, embedded into a prompt or tool call result.\n\nIt is up to the client how best to render embedded resources for the benefit\nof the LLM and/or the user.",
 "type": "object",
 "required": [
   "resource",
   "type"
 ],
 "properties": {
   "annotations": {
     "description": "Optional annotations for the client.",
     "$ref": "#/definitions/Annotations"
   },
   "resource": {
     "anyOf": [
       {
         "$ref": "#/definitions/TextResourceContents"
       },
       {
         "$ref": "#/definitions/BlobResourceContents"
       }
     ]
   },
   "type": {
     "type": "string",
     "const": "resource"
   }
 }
}

Fields
annotations: Option<Annotations>

Optional annotations for the client.

resource: EmbeddedResourceResourcetype_: String
Implementations

impl EmbeddedResource

pub fn builder() -> EmbeddedResource

impl EmbeddedResource

pub fn new(resource: impl Into<EmbeddedResourceResource>) -> Self

Creates a new EmbeddedResource with the specified resource.

This constructor initializes an embedded resource object with the given resource, setting default values for other fields.

Parameters

resource - The embedded resource

Examples

use mcp_daemon::schema::{EmbeddedResource, EmbeddedResourceResource};
let resource = EmbeddedResourceResource::new("example_resource");
let embedded_resource = EmbeddedResource::new(resource);