mcp_daemon::schema::types
Struct ProgressNotificationParams

pub struct ProgressNotificationParams {
    pub message: Option<String>,
    pub progress: f64,
    pub progress_token: ProgressToken,
    pub total: Option<f64>,
}

ProgressNotificationParams

JSON schema

{
 "type": "object",
 "required": [
   "progress",
   "progressToken"
 ],
 "properties": {
   "message": {
     "description": "An optional message describing the current progress.",
     "type": "string"
   },
   "progress": {
     "description": "The progress thus far. This should increase every time progress is made, even if the total is unknown.",
     "type": "number"
   },
   "progressToken": {
     "description": "The progress token which was given in the initial request, used to associate this notification with the request that is proceeding.",
     "$ref": "#/definitions/ProgressToken"
   },
   "total": {
     "description": "Total number of items to process (or total progress required), if known.",
     "type": "number"
   }
 }
}

Fields
message: Option<String>

An optional message describing the current progress.

progress: f64progress_token: ProgressToken

The progress token which was given in the initial request, used to associate this notification with the request that is proceeding.

total: Option<f64>
Implementations

impl ProgressNotificationParams

pub fn builder() -> ProgressNotificationParams