mcp_daemon::schema::types_ex
Struct Tag

pub struct Tag<T>(pub T);

Type representing a tag string associated with a type

This type is used when you want to output a tag string associated with a type in JSON serialization, and check if the tag string matches when deserializing.

The tag string is specified by the TAG constant of the TagData trait.

Example

use mcp_daemon::schema::{Tag, TagData};
use serde_json::json;

#[derive(Default)]
struct MyTag;

impl TagData for MyTag {
    const TAG: &'static str = "my-tag";
}

let tag = Tag(MyTag::default());
let json = serde_json::to_value(&tag).unwrap();
assert_eq!(json, json!("my-tag"));

let tag: Tag<MyTag> = serde_json::from_value(json).unwrap();

Tuple Fields
0: T